back/requirements.txt
Django==4.2.29
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==11.0.0
django-unfold==0.40.0
boto3==1.35.0
django-storages==1.14.4
